int to float correction

+++ ros_packages_full/src/RobotLeg/src/RobotLegRos.cpp
@@ -29,9 +29,9 @@ std::array<float, 3> RobotLegRos::forward_kinematics(const std::array<float, 3> 
 
 std::array<float, 3> RobotLegRos::inverse_kinematics(const std::array<float, 3> xyz_pos){
     std::array<float, 3> result_rad;
-    int x = xyz_pos[0];
-    int y = xyz_pos[1];
-    int z = xyz_pos[2];
+    float x = xyz_pos[0];
+    float y = xyz_pos[1];
+    float z = xyz_pos[2];
     result_rad[0] = atan(y/x);
     result_rad[2] = acos((pow(x-a_1_, 2) + pow(z +  h_1_, 2) - pow(a_2_, 2) - pow(a_3_, 2))/(2*a_2_*a_3_));
     result_rad[1] = atan((z +  h_1_)/(x - a_1_)) + asin(a_3_ * sin(result_rad[2])/(sqrt(pow(x - a_1_, 2) + pow(z + h_1_, 2)))); // TODO - dlaczego "-" a nie "+" ? XD
